Split Database into per-entity files
db/client.rs had grown to ~4k lines holding one impl Database block that
spanned every entity. Relocate that single impl, verbatim, into db/client/
with one file per entity (album, artist, track, release, identity, blobs,
playback); mod.rs keeps the types, constructors, shared row-mappers, and the
SQL-row helper layer the entities share. No type, signature, or behaviour
change — the only edits are the relative-path fixups the deeper location
forces (the test modules' super paths and one super::models reference).
